Refactor: move House and Room classes
