+++ src/plugins/skills/construction/room-builder.ts
@@ -1,15 +1,17 @@
 import {
     objectInteractionActionHandler
 } from '@engine/action/pipe/object-interaction.action';
-import { openHouse, Room } from '@plugins/skills/construction/house';
-import { MAP_SIZE, roomBuilderButtonMap } from '@plugins/skills/construction/con-constants';
 import { buttonActionHandler } from '@engine/action/pipe/button.action';
-import { getCurrentRoom } from '@plugins/skills/construction/util';
 import { Player } from '@engine/world/actor/player/player';
 import { Coords } from '@engine/world/position';
 import { dialogue, execute, goto } from '@engine/world/actor/dialogue';
 import { saveHouse } from './home-saver';
 import { widgets } from '@engine/config';
+
+import { getCurrentRoom } from './util';
+import { MAP_SIZE, roomBuilderButtonMap } from './con-constants';
+import { openHouse } from './house';
+import { Room } from './models/room';
 
 
 const newRoomOriention = (player: Player): number => {
